Fix datadog

+++ src/main/kotlin/xyz/gnarbot/gnar/listeners/BotListener.kt
@@ -104,7 +104,7 @@ class BotListener(private val bot: Bot) : ListenerAdapter() {
         bot.players.destroy(event.guild)
         bot.datadog.gauge("octave_bot.guilds", bot.shardManager.guildCache.size())
         bot.datadog.gauge("octave_bot.users", bot.shardManager.userCache.size())
-        bot.datadog.gauge("octave_bot.players", bot.players.size())
+        bot.datadog.gauge("octave_bot.players", bot.players.size().toLong())
         bot.datadog.incrementCounter("octave_bot.guildLeave")
     }
